Extract win information
The win screen had a bunch of logic for calculating scores using
Adagrams and the game state. This is similar to what redux calls
"selectors". Extract those "selectors" into a class so that
additional logic can be added. The extraction was done under
test, but the tests rely on Adagrams working correctly, so I'm not
merging them as-is.