docs: Higher-order reducers are not middleware
I used the term "middleware" incorrectly. Middleware is a higher-order
function applied to the `dispatch` function. The structures I had called
"middleware" in the documentation are actually higher-order
reducers. They modify the reducer, but still guarantee a state is
returned and do not produce side-effects. Middleware wrapping the
dispatcher allows for side-effects and behavior that does not produce
state modifications.